update sql, use 'Rich' display library
This modifies the SQL to use a single query to get both AlbumArtists and
TrackArtists for a given search. They come back as 1 entry per
(album, artist, track artist, playlist) pairing.

Also switched to the Rich python package for displaying output. Main
reason for switching was that tables will automatically adjust to the
terminal width.